Set prgname to fix icon in phosh (MR 13)
The icon when the app is running in phosh is missing. This is due
to an issue in gtk3[1]. This has caused problems in other apps, but
Purism did a workaround by using dbus names[2][3]. However, as
this tweaks application does not use dbus, the workaround in phosh
isn't really working. As tweaks is not a cli application, setting
prgname as the applicattion id should not really be an issue.

[1] https://gitlab.gnome.org/GNOME/gtk/-/issues/2822
[2] https://source.puri.sm/Librem5/OS-issues/-/issues/224
[3] https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/282