pmos_tweaks: datasource: also respect ~/.themes and ~/.icons (MR 19)
A user in the unofficial postmarketOS Telegram group was having issues
with their themes and icons not being detected, and the problem was that
they had installed them to these directories, which while arguably ugly
given that they pollute the home directory's top-level namespace,
clearly are used by some users. As such, let's support them in addition
to the other paths.